add EnderecoUser model
Separando endereço de usuario

+++ app/Http/Controllers/Api/CarrinhoController.php
@@ -16,7 +16,6 @@ class CarrinhoController extends Controller
         $request->validate([
             'loja_id' => 'required',
             'endereco_id' => 'required',
-            'price' => 'required'
         ]);
 
         $user = auth()->user();
@@ -25,10 +24,20 @@ class CarrinhoController extends Controller
             'user_id' => $user->id,
             'loja_id' => $request->loja_id,
             'endereco_id' => $request->endereco_id,
-            'price' => 0.00
+            'price' => 0.00,
+            'closed' => false,
         ]);
 
-        $produtos = $carrinho->produtos()->get();
+        $carrinhoProduto = $carrinho->produtos()->get();
+        $i = 0;
+        $produtos = [];
+        foreach($carrinhoProduto as $produto){
+            $produtos[$i] = [
+                'detalhe' => $produto,
+                'produto' => $produto->produto()->get()
+            ];
+            $i++;
+        }
 
         return response([
             "carrinho" => $carrinho,
@@ -42,6 +51,7 @@ class CarrinhoController extends Controller
 
         $carrinho = Carrinho::where('user_id', $user->id)
         ->where('id', $carrinho_id)
+        ->where('closed', false)
         ->first();
 
         if(empty($carrinho)){
@@ -50,20 +60,88 @@ class CarrinhoController extends Controller
             ], 404);
         }
 
-        $produtos = $carrinho->produtos()->get();
+        $carrinhoProduto = $carrinho->produtos()->get();
+        $i = 0;
+        $produtos = [];
+        foreach($carrinhoProduto as $produto){
+            $produtos[$i] = [
+                'detalhe' => $produto,
+                'produto' => $produto->produto()->get()
+            ];
+            $i++;
+        }
+        $loja = $carrinho->loja()->get();
+        $endereco = $carrinho->endereco()->get();
 
         return response([
             "carrinho" => $carrinho,
             "produtos" => $produtos,
+            "loja" => $loja,
+            "endereco" => $endereco
         ], 200);
+    }
+
+    public function updateCarrinho($carrinho_id, Request $request)
+    {
+        $request->validate([
+            'endereco_id' => 'nullable',
+            'loja_id' => 'nullable'
+        ]);
+
+        $user = auth()->user();
+        $carrinho = Carrinho::where('user_id', $user->id)
+        ->where('id', $carrinho_id)
+        ->where('closed', false)
+        ->first();
+
+        if(empty($carrinho)){
+            return response([
+                'message' => 'Carrinho não encontrado'
+            ], 404);
+        }
+
+        if($request->endereco_id){
+            $carrinho->endereco_id = $request->endereco_id;
+        }
+        if($request->loja_id){
+            $carrinho->loja_id = $request->loja_id;
+        }
+
+        $carrinho->save();
+
+        $produtos = $carrinho->produtos()->get();
+
+        foreach($produtos as $produto){
+            if($produto->loja_id != $request->loja_id){
+                $produto->delete();
+            }
+        }
+
+        $carrinhoProduto = $carrinho->produtos()->get();
+        $i = 0;
+        $produtos = [];
+        foreach($carrinhoProduto as $produto){
+            $produtos[$i] = [
+                'detalhe' => $produto,
+                'produto' => $produto->produto()->get()
+            ];
+            $i++;
+        }
+
+        return response([
+            "message" => 'Carrinho atualizado com sucesso',
+            "carrinho" => $carrinho,
+            "produtos" => $produtos,
+        ], 200);
+
     }
 
     public function insertProduto($carrinho_id, Request $request)
     {
         $request->validate([
-            'amount' => 'request|integer',
+            'amount' => 'required|integer',
             'produto_id' => 'required',
-            'note' => 'max:255'
+            'note' => 'string|max:255|nullable'
         ]);
 
         $user = auth()->user();
@@ -86,6 +164,12 @@ class CarrinhoController extends Controller
             ], 404);
         }
 
+        if($produto->loja_id != $carrinho->loja_id){
+            return response([
+                'message' => 'Produto de lojas diferentes'
+            ], 406);
+        }
+
         $carrinho_produto = CarrinhoProduto::where('carrinho_id', $carrinho_id)
         ->where('produto_id', $produto->id)
         ->first();
@@ -98,11 +182,25 @@ class CarrinhoController extends Controller
                 'note' => $request->note,
             ]);
         }else{
-            $carrinho_produto->amount = ($carrinho_produto->amount + 1);
+            $carrinho_produto->amount = ($carrinho_produto->amount + $request->amount);
             $carrinho_produto->save();
         }
 
-        $produtos = $carrinho->produtos()->get();
+        $carrinhoProduto = $carrinho->produtos()->get();
+        $i = 0;
+        $price = 0;
+        foreach($carrinhoProduto as $detalhe){
+            $produto = $detalhe->produto()->first();
+            $price = (($produto->price * $detalhe->amount) + $price);
+            $produtos[$i] = [
+                'detalhe' => $detalhe,
+                'produto' => $produto
+            ];
+            $i++;
+        }
+
+        $carrinho->price = round($price, 2);
+        $carrinho->save();
 
         return response([
             "carrinho" => $carrinho,
@@ -110,8 +208,13 @@ class CarrinhoController extends Controller
         ], 200);
     }
 
-    public function removeProduto($carrinho_id, $produto_id)
+    public function removeProduto($carrinho_id, $produto_id, Request $request)
     {
+        $request->validate([
+            'amount' => 'required|integer|nullable',
+            'delete' => 'required|boolean',
+        ]);
+        
         $user = auth()->user();
 
         $carrinho = Carrinho::where('user_id', $user->id)
@@ -134,9 +237,33 @@ class CarrinhoController extends Controller
             ], 404);
         }
 
-        $carrinho_produto->delete();
+        if($carrinho_produto->amount <= $request->amount || $request->delete == true){
+            $produto = $carrinho_produto->produto()->first();
+            $price = ($carrinho_produto->amount * $produto->price);
+            $carrinho->price = round(($carrinho->price - $price), 2);
+            $carrinho->save();
+            $carrinho_produto->delete();
+        }else{
+            $amount = ($carrinho_produto->amount - $request->amount);
+            $carrinho_produto->amount = $amount;
+            $carrinho_produto->save();
+            $produto = $carrinho_produto->produto()->first();
+            $price = $amount * $produto->price;
+            $carrinho->price = round(($carrinho->price - $price), 2);
+            $carrinho->save();
+        }
 
-        $produtos = $carrinho->produtos()->get();
+
+        $carrinhoProduto = $carrinho->produtos()->get();
+        $i = 0;
+        $produtos = [];
+        foreach($carrinhoProduto as $produto){
+            $produtos[$i] = [
+                'detalhe' => $produto,
+                'produto' => $produto->produto()->get()
+            ];
+            $i++;
+        }
 
         return response([
             "carrinho" => $carrinho,
@@ -148,7 +275,6 @@ class CarrinhoController extends Controller
     public function updateProduto($carrinho_id, $produto_id, Request $request)
     {
         $request->validate([
-            'amount' => 'request|integer',
             'note' => 'max:255'
         ]);
 
@@ -174,11 +300,19 @@ class CarrinhoController extends Controller
             ], 404);
         }
 
-        $carrinho_produto->amount = $request->amount;
         $carrinho_produto->note = $request->note;
         $carrinho_produto->save();
 
-        $produtos = $carrinho->produtos()->get();
+        $carrinhoProduto = $carrinho->produtos()->get();
+        $i = 0;
+        $produtos = [];
+        foreach($carrinhoProduto as $produto){
+            $produtos[$i] = [
+                'detalhe' => $produto,
+                'produto' => $produto->produto()->get()
+            ];
+            $i++;
+        }
 
         return response([
             "carrinho" => $carrinho,
@@ -195,8 +329,7 @@ class CarrinhoController extends Controller
         ->where('id', $carrinho_id)
         ->first();
 
-        $carrinho_produtos = CarrinhoProduto::where('carrinho_id', $carrinho->id)
-        ->get();
+        $carrinho_produtos = $carrinho->produtos()->get();
         
         if(empty($carrinho)){
             return response([
@@ -204,15 +337,18 @@ class CarrinhoController extends Controller
             ], 404);
         }
 
-        if(empty($carrinho_produto)){
+        if(empty($carrinho_produtos)){
             return response([
-                'message' => 'Carrinho não encontrado'
+                'message' => 'Carrinho Produtos não encontrado'
             ], 404);
         }
 
         foreach($carrinho_produtos as $carrinho_produto){
             $carrinho_produto->delete();
         }
+
+        $carrinho->price = 0.00;
+        $carrinho->save();
 
         return response([
             "message" => 'Carrinho limpo',
